+++ shared/src/Utils.cpp
@@ -41,7 +41,7 @@ namespace Shared {
         return min + (float) std::rand() / (float) RAND_MAX * (max - min);
     }
 
-    float Utils::radToPos(float rad)
+    double Utils::radToPos(double rad)
     {
         if (rad < 0)
             return -rad + (2 * (M_PI + rad));
